backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-groq
chromadb
sentence-transformers
PyPDF
python-dotenv
pydantic
requests
loguru
aiofiles
tqdm
logger
langchain_huggingface 
langchain-chroma
